add license info
